Added the option to read client and server settings from files. Deleted the Printer class (was used only for welcome message, maybe will be brought back later if necessary). Comments for several classes were added.

+++ src/client/Point.h
@@ -42,8 +42,8 @@ public:
    */
   virtual ~Point();
 private:
-  double x_;
-  double y_;
+  int x_;
+  int y_;
 };
 
 #endif /* POINT_H_ */
